requirements.txt
numpy
scipy
statsmodels
pandas
h5py
pysam
pybedtools
pybbi>=0.2.0
seaborn
tables
